Change filesystem access from host to xdg-pictures
Remove write access to `host` and replace with write access to `xdg-pictures` . This follows good practice to make the sand-boxing more robust. If `xdg-pictures` is not the right path or an adequate path, we can add an additional one. 

When I try to save or export a project, only the allow-listed filesystem path are shown in the file browser. Same for opening files.